Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> It's not very intuitive to have the 'incomplete' field in range lib, would it better to just keep a set of filenames with holes in the SparseImage?

Technically 'incomplete' field is not part of rangelib. rangelib only provides the 'extra' filed that can be used by the caller (similar to a "cookie").

Keeping a set of filenames with holes works. I'm thinking about using this new field also for b/64109868#comment13, which would add a different tag to mark some other files. We'll need another set of filenames for that purpose (and this "set" needs to be added to DataImage / EmptyImage as well). So seems allocating the 'extra' field in rangelib makes these a bit easier.

I was also thinking about having some "FileRangeSet" class that's derived from RangeSet to better separate the logic, which can be used as the value in file_map dict. But that requires broader changes.

> 
> Also this change makes the target-files mandatory; (so no more future ota from image files or temp directories).

Yes, that's also my initial concern. I convinced myself with the following:
a) That functionality is not ready, and will not be anytime soon;
b) With that functionality ready, we must have a tool that could read files from a given image directly (required by the signing script);
c) With b) ready, we can simply update common.GetSparseImage(), by replacing the line that "reading from target-files.zip" line with "reading from image".

So it seems having this CL doesn't make things worse... And it would allow removing the workarounds that could hide unexpected imgdiff failures.

> 
> otherwise looks good.

Patch-set: 2
Reviewer: Tao Bao <1056365@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0