Update patch set 3
Patch Set 3:

> Patch Set 3:
> I don't understand why this change fixes the issue described in the bug. You're still linking the same libc++, just a static version of it.

This is because the symbol is hidden in the shared library but the linker freely gets it from a static library.

> Is this another case where shipping a prebuilt libc++ for everyone could make things consistent?

Possibly.  But definitely due to change in libcxx.  In the prebuilts (http://aosp/c/585812), runtimes_ndk_cxx/aarch64/libFuzzer.a, which links using the libcxx in r16, does not depend on ~basic_string(), while lib64/clang/7.0/lib/linux/aarch64/libFuzzer.a, that uses toolchain/libcxx, depends on ~basic_string().  (For reference, the mangled name of the missing symbol is _ZNSt3__112basic_stringIcNS_11char_traitsIcEENS_9allocatorIcEEED2Ev).

Patch-set: 3
Reviewer: Pirama Arumuga Nainar <1058860@85c56323-6fa9-3386-8a01-6480fb634889>