Update patch set 1
Patch Set 1: Code-Review-1

Looks like

1) I'm pulling shared libraries into general-tests.zip so probably don't want to do that and filter it to apks/gtests.

2) Looks like gcmtests-oldkey doesn't build properly even without my change... so now we try to build it and it fails... sigh... https://cs.corp.google.com/android/vendor/google/tests/GCM/Android.mk?dr=C

Not sure what to do about that one as it's an APK Instrumentation test so technically it should be runnable.

From that cs link "# Make gcmtests-oldkey.apk for Android prior to ICS." So why are we building it now? Is this related to the fact that test_suites isn't building Android but additional artifacts?

I'm trying to do `make general-tests` now from my checkout to see if it hits this error.

Patch-set: 1
Reviewer: Simran Basi <1083173@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Presubmit-Ready=0