Change the naming policy of system_$(VER)
The name of system_$(VER) prebuilt file in droiddoc.mk and
java_common.mk was changed to sdk_v$(LOCAL_SDK_VERSION). But
system_$(VER) in apidiff.mk is not changed, since apidiff.mk
file is not in AOSP. So this modify is added and the modify
for core_current is added.

Bug: 72031391
Test: build & run on taimen
Change-Id: Ie227f4c6911fc80484f12b3f1cfa6472cb605d4b