Clarify error message for bootclasspath whitelist.
Before this CL, the build could fail with error messages such as:

Error: out/target/common/obj/JAVA_LIBRARIES/core-oj_intermediates/classes.jar
contains class file jdk/internal/HotSpotIntrinsicCandidate.class, which is not
in the whitelist

This error message was only moderately helpful because it left a few
questions unanswered or misled ("Whitelist for what?", "Where does the
whitelist live?", "Is it a whitelist of class files or of packages?").

This CL clarifies that:
 - it's a whitelist of packages allowed on the bootclasspath,
 - where it lives (currently
   build/make/core/tasks/check_boot_jars/package_whitelist.txt)
which makes the error message more actionable.

Test: manually checked that the error message now looks okay.
Bug: 17434570

Change-Id: I2f52a5e2eb532bc4945bedf9811de5857f67a9a3