Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Have you verified oc-mr1-dev-plus-aosp & pi-dev-plus-aosp too? (There's a build_test target, so you can run that via forrest)
> > 
> > No, I haven't done that yet. I'm planning on merging both of these tomorrow morning LON time,  but I'll check those trees before merging this one (and then do pi-dev..).
> > 
> > Never used forrest.. I guess go/android-forrest the right place to start?
> 
> Yep, based on my command line history, something like:
> 
> forrest --gerrit_hostname android --force_cherry_pick build th:cl:635051:build_test:git_oc-mr1-dev-plus-aosp

Thanks for that! Nice tool. We can't enforce this in oc-mr1-dev-plus-aosp, which makes sense. I plan on submitting this and then reverting it in that branch (not sure how else to best submit it).

pi-dev-plus-aosp needs ag/3696341, then I'll submit this.

Patch-set: 3
Reviewer: Anton Hansson <1060959@85c56323-6fa9-3386-8a01-6480fb634889>