Update patch set 1
Patch Set 1:

I'm not sure this is the right approach, but we do need an option which allows device manufacturers to preload APK in /system/priv-app having compressed dex. This is very important for the devices (1) upgrading from O to P but not having enough free space in system partition, (2) shipping Android Go edition which targets low-tier models.

Patch-set: 1