Allow radio to read emulator network properties
A recent change added SELinux labels to the properties used by the DHCP
software and RIL to configure the network. Unfortunately that change
didn't give RIL the permissions needed to read those properties which
broke radio networking for the emulator. Fix it by allowing radio
related code to read the properties again.

BUG: 76211046
Test: Run emulator with -feature -Wifi and verify network connectivity
Change-Id: I7663a6598e2d501ee8336b2dae5fd78ff4ff69bc