Change the default compiler filter for install to speed-profile
The install filter is speed-profile in order to enable the use of profiles
from the dex metadata files. Note that if a profile is not provided or if
it is empty speed-profile is equivalent to (quicken + empty app image).

Test: build & install an app
Bug: b/30934496

(cherry picked from commit b5dadc3a18c94625d3d473200f146cea55ba91c0)

Merged-Id: I6671792b6c846c570a45d3c6a9b9677bcc424b17
Change-Id: I6eb81934a0b7cbaec4ff4471bf465cd40967d40b