Add 'honggfuzz' and 'fuzz' to all ASAN / coverage builds
'hongfuzz' is a the interface to all honggfuzz fuzzers.
Unlike libFuzzer, where the entry point *is* the fuzzer binary,
honggfuzz uses a separate binary to drive fuzzing.

'fuzz' is a wrapper script that performs some device-side setup
and sanity checks for fuzzing sessions, while providing a simple
interface to all user-space fuzzers.

Change-Id: Ia55d517e9cb1b7bfcd7d29b137b691e74398ca9e