Disable warnings for unknown modules in package-modules.mk
We've had these warnings for a while, but we've still got a lot of
warnings that people have been trained to ignore. Lets just silence them
for now.

The Merged-In is just a random change from AOSP that isn't in pi-dev, to
stop this from being merged to master, where I'm just fixing the
warnings.

Bug: 77883700
Test: m  <fewer warnings>
Change-Id: I00d8173a527760d62dca5837bcc86d5f03b87e83
Merged-In: I78ca65e6f0c81f09e7da848eda797b3a8f97a521