Apply CFI include paths to all product configs.
The current VNDK ABI snapshot expects that CFI is disabled for
components that either in the include paths (from cfi-common.mk) OR
enabled directly in the makefile/blueprint. However, on non-arm64
architectures, the build system was only disabling CFI for vendor
components in the include paths.

This topic corrects it by (a) removing the include paths check to disable
CFI for vendor variants, and (b) inheriting the CFI include paths in
all product configs to ensure that when we update the ABI we're able
to move to relying exclusively on (PRODUCT_)CFI_INCLUDE_PATHS.

Bug: 66301104
Test: ABI matches for all architectures.
Change-Id: I4d845cd9925354d565e06020c479cc3f7ab28f61