Update patch set 1
Patch Set 1: Presubmit-Ready+1

I'm not sure if you guys have a better solution. You can see from my commit message (and looking at this dare I say?, messy, file) that this still contains a lot to be desired, but if we can get this improvement in, or if you can express a goal that gets us closer to a clean world, then I'm happy to try that as well.

Patch-set: 1
Reviewer: Steven Moreland <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1