Update patch set 1
Patch Set 1: Verified+1 Autosubmit+1

> Patch Set 1: Code-Review+2
> 
> This is strictly better than what we have, so LGTM.  We do have the finder in soong_ui that has a database of all the Android.bp and top-level Android.mk files, so we could theoretically do something even more accurate inside soong_ui.

Yeah, this is the kind of thing I was hoping we could take advantage of by "we should move more of the complexity in the build system itself so that it can intelligently find the first parent directory which contains a makefile in a subdirectory", but I don't know too many of the details.

Thanks!

Patch-set: 1
Reviewer: Steven Moreland <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Autosubmit=+1
Label: Presubmit-Ready=0