Update patch set 3
Patch Set 3: Presubmit-Verified+1

=== Finished presubmit run: L32300000182247680 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/700050?revision=3&workplanId=L32300000182247680




TreeHugger finished with: 20 passed.







WARNING: The tip-of-tree(4839221) for git_master/gce_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4839221) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4839221) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4839221) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4839221) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





This is pure great!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 3
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0
Label: Presubmit-Verified=+1