check-vndk-list is run even when VNDK is not enabled
check-vndk-list has not been run when VNDK is not enabled (i.e.
BOARD_VNDK_VERSION is unset). This was problematic because there is a
chance that a CL which passed pre-submit tests can fail the post-submit
tests when the targets selected for the pre-submit tests are all without
VNDK enforcement.

In order to prevent the case, check-vndk-list is executed regardless of
whether BOARD_VNDK_VERSION is set or not.

Bug: 110424280
Test: m -j on sailfish
Test: m -j on sailfish fails after adding an arbitrary library to VNDK
Change-Id: Ia7427be2a297940a9e9c9f46cb058245596171c9