Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > I'd rather figure out how to make this automatic.  Otherwise adding a test in a directory that already has a library would require adding this to the library.
> > 
> > Here is the https://cs.corp.google.com/android/art/tools/ahat/Android.mk?dr&g=0&l=33 in question. The situation is that the executable is needed by other tests for execution but also has a test located right next to it.
> > 
> > Any thoughts/initial direction on how to automate this?
> > 
> > One thought I have is that the test config can only be applied to a select white-list of classes, thought I suspect we might need to white-list executables for other use cases...
> > 
> > A second idea is to autogenerate HostTest configs and remove the test xml file that is here already and apply LOCAL_DISABLE_AUTO_GENERATE_TEST_CONFIG to ahat, but that has a similar issue to what you called out for this CL.
> > 
> > 
> > Alternatively, a larger fix/refactor, would be to revisit the usage of L_C_S and maybe handle test dependencies in a different manner...
> 
> The ideal solution would be to add LOCAL_REQUIRED_MODULES := ahat to the tests that need it, and then make sure that required modules get packaged into the zip.

So we talked about doing that a number of times going back to June 2017 - my uncommited CL: https://android-review.googlesource.com/c/platform/build/+/416916 - but the ideal path forward DanW and I were thinking to eventually get to would be that tests would have a primary install location under out/target/../testcases and out/host/../testcases and then we can use LOCAL_REQUIRED_MODULES and it would just work.

My unimplemented design doc for that approach:https://docs.google.com/document/d/1jXEnm27OvqjIto6mruduqYUh1H1ifWtB3_sF9lbM4dM/edit (DanW and I agreed we could not make the change in aosp and break the old test tools unless we have atest in aosp)

But this case is different b/c ahat is not a test, it's a tool that tests need to run. So that design would actually not work for this specific case b/c ahat by default installs into out/host/linux-x86/bin


I don't remember the exact reasoning why Dan determined that CL would not be possible but we discussed it in person (something to do with we should not create new deps in main.mk and the fact I would not be able to figure out what files to copy in base_rules.mk).


If we have a new way to make LOCAL_REQUIRED_MODULES work as expected (i.e. install in all locations) that would be great. Note we are running: `make <dependent target>' not 'make <suite>'

Patch-set: 1