Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I know Bill added a duplicate variable, we should update that code to use this new variable.

Hi Simran,

Not sure if you want to change this part like this way.
https://googleplex-android-review.git.corp.google.com/c/platform/build/+/4738622

But it will not work.
The config's content will be like this if building hello_world_test.
   <option name="module-name" value="xts-xml-generator" />

Here's the comment from Dan W before.
The contents of Make rules are not evaluated immediately, they're evaluated at the end of the build, so you're getting the last value of LOCAL_MODULE. By convention, PRIVATE_* variables are target-specific variables, which can be local to a particular target.

I think it still needed to duplicate LOCAL_MODULE to PRIVATE for test config autogen process.

Patch-set: 2
Reviewer: Gerrit User 1308044 <1308044@85c56323-6fa9-3386-8a01-6480fb634889>