Add additional OWNERS for entries to the VNDK.
The target/product/vndk/current.txt file was added in order
to have a global check on what is included in the VNDK. This
for instance adds as an extra check to make sure that a
device-specific module doesn't end up in this global set.

However, because modifying the VNDK is relatively common, this
puts undue burden on the build team to conduct reviews. For that
reason, I'm opening it up with additional reviewers.

Bug: N/A
Test: N/A
Change-Id: I9827260bde8ec757be363fc153b718d9c0b68f25