Add recovery-persist to base_system
We moved the recovery metric parsing to the native code, which executes
on the next normal build and stays under the system partition. Also,
the recovery-persist binary needs to be installed on both A/B and
non-A/B devices. Therefore, adds recovery-persist to base_system to
avoid breaking the mainline targets.

Bug: 114278989
Test: build mainline target
Change-Id: I251ef862d9342b0dc308223ee6c48849f098bb1e