Experiment to be able to change the navigation bar thickness (2/2)
Use resource overlays to change the 2 dimension navigation_bar_height
and navigation_bar_frame_height to specify how large the navigation bar
is and how large its buttons are (also for width). There are two
overlays: first to slim the navigation bar to 75% and the other to have
floating navigation buttons.

go/navbar-experiments-doc

Test: manual
Bug: 113952590
Change-Id: Ibd93097e223cbf0de18011064b3da38b6d069d24