Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > There isn't a temporary build directory for these files, because we don't really want to be modifying PRODUCT_COPY_FILES at all, especially not during installation which has to be redone after m installclean.
> 
> Can't we just `rm $@.unaligned`, similar to other rules?

Other rules are operating in their own directory, not in the shared $OUT/system directory.  We shouldn't be creating and removing files in there, that can confuse tools that are operating on the system directory.

> > The ideal solution would be to replace the modification with a check that forbids uncompressed dex files in jars, or forbids jars altogether.  Then modify all the vendor files to use a prebuilt module instead of PRODUCT_COPY_FILES for their jars.
> 
> This has much more impact as it involves lots of rules/prebuilts not owned by us, plus the vendors will have to update their tools to generate jars with uncompressed dex.  I don't see clear benefit comparing to the build system itself changing the compression ratio?

The build system can still be changing the compression ration.  If the modules are BUILD_PREBUILT instead of PRODUCT_COPY_FILES then we can uncompress the dex files in the intermediates directory before installing them with a plain copy, which is how the build is supposed to work.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0