Update patch set 4
Patch Set 4:

(2 comments)

My main concern with this is that it makes it easy for changes to break only a few obscure builds.  If we need this, we need to ensure it is enabled in something that is in the global presubmit checks.  I'd also like a plan to get it enabled everywhere.  Dan's been doing a lot of changes like that, with the general flow of enable in a few, then explicitly disable in ones that have problems, and then flip the default to enabled.

Patch-set: 4
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>