Update patch set 1
Patch Set 1:

> Can't we just have a "filename" or "output_name" property on java_sdk_library that allows us to override the default jar name? i.e. instead of outputting foo.impl.jar, when filename is specified it uses the value and allows us to stop the artifact changing where it affects things on device?

We can change "filename" or "output_name". So we can use the core-simple.jar instead of core-simple.impl.jar. But most of name changes are module name, not file name. And we can not use the "core-simple" module because the "core-simple" is already used by the java_sdk_library module. In conclusion, I think changing only the file name would be more confusing.

Then we have a way to specify the stubs module name and impl module name. If so, we can not use linking with the appropriate stubs library according to the sdk_version.

So, I think creating another module for the core library is better than to use java_sdk_library because java_sdk_library is designed for uses library and java_sdk_library has to be changed a lot.

Neil, what do you think about it?

Patch-set: 1
Reviewer: Gerrit User 1187124 <1187124@85c56323-6fa9-3386-8a01-6480fb634889>