Add rss_hmw_reset to system partition
rss_hwm_reset is binary that reset RSS high-water mark counters for all
currently running processes. It runs in a separate process because it
needs dac_override capability.

The binary is not a persistent service. It is invoked from a system
service by flipping sys.rss_hwm_reset.on property.

Bug: 119603799
Test: adb shell ls /system/bin
Change-Id: Iee0b54ec6486d8f23048819493f6d710db3feff6