Update patch set 3
Patch Set 3:

Formatting the product partition sounds good to me.

Is the point of the unmount such that the there is a /product symlink pointing to GSI and the unmount will allow that content to be used in place of the vendor's product.img?  There's still sort of the same issue here that any init scripts in /system/product/... won't be executed.  Though that's something that I guess can be taken into consideration when making /product.

Otherwise, if you're not trying to use such a symlink, is there even a point in unmounting /product?  Formatting that partition should be sufficient.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>