Update patch set 2
Patch Set 2: -Autosubmit

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> I know I had commented on the bug that we may need separate product variables for each of the different partitions, but it seems like this is basically saying use the one input and filter the contents per output partition.  So then vendors would have a single config.fs?  Seems like I could do the same thing for AIDs and just filter out the various classes per partition?  (Given that I'll need to make new classes for each of the partitions).

The one source is merely a convenience, we can merge them and explode them at will has been my expectation. It deals with our expectations for the other partitions, and we provide an override. Saves a lot of hard-coded defaults in the hands of the partners.

AIDs are an orthogonal sliced out set of OEM, so at least that part is perhaps easier. There is no overlap, per-se. easy-peasy.

On the other hand, for fs_config there _is_ overlap in expectations. The build.prop set we see is just the tip of the iceberg. We should have an expectation/default for product and product_services, for instance system/bin, product/bin and product_services/bin and the default for their contents should all be the same.

The overlap needs to be removed in the Treble world though. Perhaps we should have an Android default config.fs for each partition. (config_system.fs, config_vendor.fs, config_product.fs ...). We can validate each, as well as the partner supplied overrides to be very specific to their ortogonal directory trees. They can be mixed and separated out at will during the build processes with that enforced validation. Mixing them is necessary for the case where they are bundled (eg: vendor _is_ under /system/vendor), so all should be searched by the tool(s).

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=0
Label: Presubmit-Ready=0