Do not install unused arch variant of audio libs
PRODUCT_PACKAGES installs the primary arch variant of executables
and both the primary and secondary arch variant of shared libraries.
It also installs the dependencies of each files.

A lot of audio libraries were listed in PRODUCT_PACKAGES although
they are dependencies of audioserver.
Thus those libraries where installed in both archs whereas
only the arch of the audioserver is needed.

Thus remove all audio libraries from PRODUCT_PACKAGES and have them
installed automatically as a dependency of audioserver in the arch
variant of audioserver.

That should also save ~3MB of space on the system partition.

Test: m -j; du -a $ANDROID_PRODUCT_OUT/system > before
      # apply this patch
      rm -r $ANDROID_PRODUCT_OUT
      m -j; du -a $ANDROID_PRODUCT_OUT/system > after
      meld before after
      # lib64 is not changed, lib32 does not have audio libs
Signed-off-by: Kevin Rocard <krocard@google.com>

Change-Id: Icbcf4c92d60819ca2f9c14ba0aec043584ed081f