Remove km_MH locale from mainline
This doesn't appear to be a real locale. The interpretation would be
khmer spoken in the Marshall Islands, which doesn't appear to be a thing.
The locale doesn't appear in https://lh.2xlibre.net/locales/ or
frameworks/base/core/res/res/values/locale_config.xml.

It might have been a typo when it was introduced in internal change
I78d91916fdbb04502466c26bf7d160fe679aaa31.

Test: make
Change-Id: Id1deebdd8ad2a5601c15b84e010be766b3a951e9