Move overridden AOSP modules out of *_system mks
All these modules are to be moved to /product, as they are not
part of the core system image. Move them out of the _system mks
as a start, and temporarily whitelist them so that they can be
moved iteratively.

Bug: 122353713
Test: make
Change-Id: If658e6b4565924cd915a183400c5b2b32984eca4
Merged-In: If658e6b4565924cd915a183400c5b2b32984eca4