Enable mini-debug-info for JITed methods by default.
Generate and store minimum amount of information necessary to
create backtraces in native tools (for crashes and profilers).

The data is compressed and takes <5% relative to JIT code size
(which is less than recent stack map savings; so win overall).

Averages for non-compiled run of maps which filled JIT code cache to ~4MB:
Q: code:1.55kb/method data:0.738kb/method debug:0.064kb/method(*this CL*)
P: code:1.62kb/method data:0.956kb/method debug:0.835kb/method(if enabled)

I also measured the P to Q on-disk stack maps savings in CL/762841 as ~28%
(not all of that is applicable to JITed code, so that is an overestimate).

Bug: 111350693
Test: Device boots, mini-debug-info is generated.
Change-Id: I95883bd4f2d5c2308149ab0ba641b0a703c7f81b