Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 1: Presubmit-Ready+1
> 
> +steven was android.hidl.base@1.0 never a module? Not sure why I added it in the first place

android.hidl.base@1.0 is a HIDL package that exists, and by default all HIDL packages have a library with their package name. However, because this interface is special, and because of some circular dependencies (and Android O craziness, really this dependency is kind of a mess), this library is built into libhidltransport, and so to prevent OD-R violation, the build system doesn't create a library for this package (and things just include libhidltransport). It should be safe to remove.

Funny story, some devices still copy/use this for I don't know what reason...

Patch-set: 2
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1