Update patch set 3
Patch Set 3:

For better or worse, this is an established API -- and while Android.mk files have been using this less due to the Soong conversion, I expect a number of vendor scripts to be using this still (though I don't know how many need to deal with host modules).

I think we'd at least need to split PRODUCT_PACKAGES for host vs target first, so that we can prepend the host_ prefix. I've been preparing for this, but haven't started yet. Then there's still all the other places that specify lists of packages.

Patch-set: 3
CC: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>